refactor help/describe screen to be in the table view mod
