+++ src/rust/kanto-tui/src/ui/mod.rs
@@ -1,126 +1,13 @@
-use crate::cm_types::Container;
 use cursive::{traits::*, Cursive};
 
 use super::{
-    cm_types,
-    kantui_config::{AppConfig, ALT_REPR, CTRL_REPR},
-    try_best, KantoRequest, KantoResponse, RequestPriority, Result,
+    kantui_config::AppConfig, try_best, KantoRequest, KantoResponse, RequestPriority, Result,
 };
 use async_priority_channel::{Receiver, Sender};
-use cursive::views::{Dialog, OnEventView, TextView};
+use cursive::views::Dialog;
 
 pub mod containers_table_view;
 use containers_table_view as table;
-
-fn host_config_description(host_config: &cm_types::HostConfig) -> String {
-    format!(
-        r"Network mode: {}
-    Port Mappings: {:#?}
-    Privileged: {},
-    Devices: {:#?}",
-        host_config.network_mode,
-        host_config.port_mappings,
-        host_config.privileged,
-        host_config.devices
-    )
-}
-
-fn describe_screen(siv: &mut Cursive, c: Container) {
-    use cursive::event::Key::Esc;
-    let cnt_description = format!(
-        r"
-    General
-    ========================
-    ID: {}
-    Name: {}
-    Container hostname: {}
-    Image: {}
-    State: {}
-
-    Host Config
-    ========================
-    {}
-
-    Other
-    =======================
-    Mounts: {:#?}
-    ",
-        c.id,
-        c.name,
-        c.host_name,
-        c.image.map_or("N/A".to_string(), |image| image.name),
-        c.state.map_or("N/A".to_string(), |state| format!(
-            "{} (Exit code: {})",
-            state.status, state.exit_code
-        )),
-        c.host_config
-            .map_or("N/A".to_string(), |config| host_config_description(&config)),
-        c.mounts
-    );
-
-    let describe_view = Dialog::around(TextView::new(cnt_description))
-        .title("Container Description")
-        .button("Ok (Esc)", |s| try_best(s.pop_layer()))
-        .scrollable();
-
-    let describe_events_handler =
-        OnEventView::new(describe_view).on_event(Esc, |s| try_best(s.pop_layer()));
-
-    siv.add_layer(describe_events_handler);
-}
-
-fn help_screen(siv: &mut Cursive, config: AppConfig) {
-    use cursive::event::Key::Esc;
-    let help_string = format!(
-        r"
-    You can use either the arrow keys/Tab/Enter (keyboard) 
-    or the mouse (if your terminal supports mouse events) 
-    to select a container from the list.
-
-    UI Button/Keyboard Shortcut => Function
-    ==================================================
-    {}/{} => To Start the currently selected container
-    {}/{} => To Stop the currently selected container
-    {}/{} => To Remove the currently selected container
-    {}/{} => To Get Logs for the currently selected container
-    {}/{} => To Redeploy all container manifests
-    {}/{} => To Display this help screen
-    {}/{} => To Exit Kantui
-    ==================================================
-
-    Legend:
-    ============================
-    {CTRL_REPR}<key> = Ctrl+<key>
-    {ALT_REPR}<key> = Alt+<key>
-    ",
-        config.keyconfig.start_btn_name,
-        config.keyconfig.start_kbd_key,
-        config.keyconfig.stop_btn_name,
-        config.keyconfig.stop_kbd_key,
-        config.keyconfig.remove_btn_name,
-        config.keyconfig.remove_kbd_key,
-        config.keyconfig.logs_btn_name,
-        config.keyconfig.logs_kbd_key,
-        config.keyconfig.redeploy_btn_name,
-        config.keyconfig.redeploy_kbd_key,
-        config.keyconfig.help_btn_name,
-        config.keyconfig.help_kbd_key,
-        config.keyconfig.quit_btn_name,
-        config.keyconfig.quit_kbd_key
-    );
-
-    let help_view = Dialog::around(TextView::new(help_string))
-        .title("Help")
-        .button("Ok (Esc)", |s| try_best(s.pop_layer()))
-        .scrollable()
-        .scroll_y(true)
-        .scroll_x(true);
-
-    let help_events_handler =
-        OnEventView::new(help_view).on_event(Esc, |s| try_best(s.pop_layer()));
-
-    siv.add_layer(help_events_handler);
-}
 
 /// Setup the user interface and start the UI thread
 pub fn run(
@@ -167,7 +54,8 @@ pub fn run(
     let redeploy_cb = enclose::enclose!((tx_requests) move |_s: &mut Cursive| {
         try_best(tx_requests.try_send(KantoRequest::Redeploy, RequestPriority::Normal));
     });
-    let help_cb = enclose::enclose!((config) move |s: &mut Cursive| help_screen(s, config.clone()));
+    let help_cb =
+        enclose::enclose!((config) move |s: &mut Cursive| table::help_screen(s, config.clone()));
 
     siv.add_fullscreen_layer(
         Dialog::around(table.with_name(table::TABLE_IDENTIFIER).full_screen())
@@ -203,7 +91,9 @@ pub fn run(
             match resp {
                 KantoResponse::ListContainers(list) => table::update_table_items(s, list),
                 KantoResponse::GetLogs(logs) => table::show_logs_view(s, logs),
-                KantoResponse::GetFullContainerState(container) => describe_screen(s, container),
+                KantoResponse::GetFullContainerState(container) => {
+                    table::describe_screen(s, container)
+                }
             }
         }
     });
